Cleaned up the window show()/hide() code, made new virtual functions create() and destroy(). I was unable to remove the overloading of show() and hide() from Fl_Window, as this would break too much stuff.  But the code is much cleaner now. I also added some new methods on windows which should be much clearer than the old modal stuff:
  Fl_Window::show(parent window)

    Shows the window as a child of the given parent (which must already
    be shown).  This replaces non_modal()

  Fl_Window::exec(parent window = 0)

    Shows the window as a child of the parent (most recent window that
    got an event is used if 0 is passed), and waits until the window is
    dismissed.  Returns window's value() (which could be set by an OK
    button).  This can replace modal().

Both non_modal() and modal() are retained for back compatability.

I also cleaned up the Fl_Gl_Window some.


git-svn-id: file:///fltk/svn/trunk@715 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ FL/gl.h
@@ -1,5 +1,5 @@
 //
-// "$Id: gl.h,v 1.7 1999/08/16 07:30:54 bill Exp $"
+// "$Id: gl.h,v 1.8 1999/09/14 07:17:16 bill Exp $"
 //
 // OpenGL header file for the Fast Light Tool Kit (FLTK).
 //
@@ -61,9 +61,9 @@ FL_EXPORT double gl_width(uchar);
 FL_EXPORT void gl_draw(const char*);
 FL_EXPORT void gl_draw(const char*, int n);
 FL_EXPORT void gl_draw(const char*, int x, int y);
-FL_EXPORT void gl_draw(const char*, float x, float y);
+FL_EXPORT void gl_draw(const char*, double x, double y);
 FL_EXPORT void gl_draw(const char*, int n, int x, int y);
-FL_EXPORT void gl_draw(const char*, int n, float x, float y);
+FL_EXPORT void gl_draw(const char*, int n, double x, double y);
 FL_EXPORT void gl_draw(const char*, int x, int y, int w, int h, Fl_Flags);
 FL_EXPORT void gl_measure(const char*, int& x, int& y);
 
@@ -72,5 +72,5 @@ FL_EXPORT void gl_draw_image(const uchar *, int x,int y,int w,int h, int d=3, in
 #endif
 
 //
-// End of "$Id: gl.h,v 1.7 1999/08/16 07:30:54 bill Exp $".
+// End of "$Id: gl.h,v 1.8 1999/09/14 07:17:16 bill Exp $".
 //
